Adding option to order by DVD
This is an important option for TV Shows to allow for returning the results in DVD order instead of Airdate Order. Why is this important? Because many people are ripping their TV shows and using SickBeard to manage them and automatically name and process them.

The change here allows you to add shows like "Batman: The Animated Series" in a sane fashion and have episodes that should have aired in a different order still work.

Adding database upgrade